Unrolled Loops in multiseedSearchWorkerPaired
If a loop was constant and didn't use `paired` then it wasn't touched,
  trivial to change anytime.
A loop with many continues and break wasn't changed yet. Too many to do
quickly.
A loop with the variable nSeedRounds as control was changed since
nSeedRounds is constant across the entire program.

Code needs to be cleaned up, but logic has been tested and gives correct
output for paired. Unpaired was untouched.