Paired unrolled, scope of variables changed
Trivial loops kept the same
Beweare of recursively unrolled loops make the code look much larger,
but it is unrolled. Example is the inistializing and searching of seeds
appearing 4 times after unrolling it in a loop and then again in the
reseeding outerloop.

Missed a loop from the previous commit. Unrolled that as well.

Before most unrolled still kept their {} braces for scope, but now those
were removed wherever possible for a cleaner code.